+++ sv-tools/src/main/java/com/hartwig/hmftools/svtools/fusion_likelihood/GeneRangeData.java
@@ -27,11 +27,11 @@ public class GeneRangeData
     private Boolean mStreamUpOnly; // if a gene is forced to only be a 5' or 3' partner
 
     // maps from the DEL or DUP bucket length array index to overlap count
-    private Map<Integer,Long> mDelFusionBaseCounts;
-    private Map<Integer,Long> mDupFusionBaseCounts;
+    private Map<Integer,Integer> mDelFusionBaseCounts;
+    private Map<Integer,Integer> mDupFusionBaseCounts;
 
-    private long[] mBaseOverlapCountDownstream;
-    private long[] mBaseOverlapCountUpstream;
+    private int[] mBaseOverlapCountDownstream;
+    private int[] mBaseOverlapCountUpstream;
 
     public static final int NON_PROX_TYPE_SHORT_INV = 0;
     public static final int NON_PROX_TYPE_MEDIUM_INV = 1;
@@ -53,8 +53,8 @@ public class GeneRangeData
         mDelFusionBaseCounts = Maps.newHashMap();
         mDupFusionBaseCounts = Maps.newHashMap();
 
-        mBaseOverlapCountUpstream = new long[NON_PROX_TYPE_MAX];
-        mBaseOverlapCountDownstream = new long[NON_PROX_TYPE_MAX];
+        mBaseOverlapCountUpstream = new int[NON_PROX_TYPE_MAX];
+        mBaseOverlapCountDownstream = new int[NON_PROX_TYPE_MAX];
     }
 
     public final List<GenePhaseRegion> getPhaseRegions() { return mPhaseRegions; }
@@ -68,32 +68,32 @@ public class GeneRangeData
     public boolean isOnlyDownstreamPartner() { return mStreamUpOnly != null ? !mStreamUpOnly : false; }
     public void setRestrictedStream(Boolean stream) { mStreamUpOnly = stream; }
 
-    public Map<Integer,Long> getDelFusionBaseCounts() { return mDelFusionBaseCounts; }
-    public Map<Integer,Long> getDupFusionBaseCounts() { return mDupFusionBaseCounts; }
+    public Map<Integer,Integer> getDelFusionBaseCounts() { return mDelFusionBaseCounts; }
+    public Map<Integer,Integer> getDupFusionBaseCounts() { return mDupFusionBaseCounts; }
 
     public boolean hasCodingTranscripts()
     {
         return mPhaseRegions.stream().anyMatch(x -> x.Phase != PHASE_NON_CODING);
     }
 
-    public long getBaseOverlapCountUpstream(int type) { return mBaseOverlapCountUpstream[type]; }
-    public void addBaseOverlapCountUpstream(int type, long count) { mBaseOverlapCountUpstream[type] += count; }
-    public long getBaseOverlapCountDownstream(int type) { return mBaseOverlapCountDownstream[type]; }
-    public void addBaseOverlapCountDownstream(int type, long count) { mBaseOverlapCountDownstream[type] += count; }
+    public int getBaseOverlapCountUpstream(int type) { return mBaseOverlapCountUpstream[type]; }
+    public void addBaseOverlapCountUpstream(int type, int count) { mBaseOverlapCountUpstream[type] += count; }
+    public int getBaseOverlapCountDownstream(int type) { return mBaseOverlapCountDownstream[type]; }
+    public void addBaseOverlapCountDownstream(int type, int count) { mBaseOverlapCountDownstream[type] += count; }
 
     public void clearOverlapCounts()
     {
         mDelFusionBaseCounts.clear();
         mDupFusionBaseCounts.clear();
-        mBaseOverlapCountUpstream = new long[NON_PROX_TYPE_MAX];
-        mBaseOverlapCountDownstream = new long[NON_PROX_TYPE_MAX];
+        mBaseOverlapCountUpstream = new int[NON_PROX_TYPE_MAX];
+        mBaseOverlapCountDownstream = new int[NON_PROX_TYPE_MAX];
     }
 
-    public long phasedRegionTotal()
+    public int phasedRegionTotal()
     {
         final boolean[] codingPhases = {false, false, true, true, true};
 
-        long codingBases = mPhaseRegions.stream()
+        int codingBases = (int)mPhaseRegions.stream()
                 .filter(x -> x.hasAnyPhaseMatch(codingPhases))
                 .mapToLong(x -> x.length())
                 .sum();
